Fix several compile errors and warnings
  - errors due to missing headers
  - errors due to _GLIBCXX_STD being undefined
  - warnings due to missing return values
  - warnings due to uninitialized variables
  - warnings due to member initialization orderings
  - warnings due to signed-unsigned comparisons
  - warnings due to unused variables
  - warning due to redefined macro

+++ src/examples/slater_demo.cc
@@ -131,7 +131,7 @@ int main()
   else
   {
     cerr << "I found " << d << " different occupied orbitals." << endl;
-    for (size_t k = 0; k < d; ++k)
+    for (size_t k = 0; k < static_cast<size_t>(d); ++k)
     {
       cerr << "alpha[" << k << "] = " << alpha[k] << ", p[" << k << "] = " << p[k] << endl;
       cerr << "beta[" << k << "]  = " << beta[k]  << ", q[" << k << "] = " << q[k] << endl;
